lint: [move-only] Move text related lints to text_format.rs
